+++ Src/DlgAbout/DlgAbout.h
@@ -83,6 +83,7 @@ public:
 public:
     QString m_szAppName;
     QImage m_AppIcon;
+    //! The application version, it is diffent RabbitCommon::CTools::Version()
     QString m_szVersion;
     QString m_szVersionRevision;
     QString m_szVersionRevisionUrl;