+++ Src/Resource/Translations/RabbitCommon_zh_CN.ts
@@ -128,91 +128,103 @@
     </message>
     <message>
         <location filename="../../DlgAbout/DlgAbout.cpp" line="409"/>
+        <source>=========== RabbitCommon ========
+</source>
+        <translation>=========== 玉兔公共库 ========
+</translation>
+    </message>
+    <message>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="410"/>
+        <source>RabbitCommon version:</source>
+        <translation>玉兔公共库版本号：</translation>
+    </message>
+    <message>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="412"/>
         <source>============== Qt ===========
 </source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="410"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="413"/>
         <source>Qt runtime version: </source>
         <translation>Qt 运行版本：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="411"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="414"/>
         <source>Qt compile version: </source>
         <translation>Qt 编译版本：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="412"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="415"/>
         <source>Qt library version: </source>
         <translation>Qt 库版本：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="413"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="416"/>
         <source>Locale: </source>
         <translation>语言：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="415"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="418"/>
         <source>============== OS ===========
 </source>
         <translation>========= 操作系统 =========
 </translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="416"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="419"/>
         <source>OS: </source>
         <translation>操作系统：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="417"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="420"/>
         <source>Kernel type: </source>
         <translation>内核类型：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="418"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="421"/>
         <source>Kernel version: </source>
         <translation>内核版本：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="420"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="423"/>
         <source>Boot Id: </source>
         <translation>启动 ID：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="421"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="424"/>
         <source>Build ABI: </source>
         <translation>编译 ABI：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="422"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="425"/>
         <source>CPU: </source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="423"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="426"/>
         <source>Build CPU: </source>
         <translation>编译时 CPU：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="425"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="428"/>
         <source>============= Host ==========
 </source>
         <translation>============ 主机 =========
 </translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="426"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="429"/>
         <source>Host name: </source>
         <translation>主机名：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="427"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="430"/>
         <source>Domain name: </source>
         <translation>域名：</translation>
     </message>
     <message>
-        <location filename="../../DlgAbout/DlgAbout.cpp" line="431"/>
+        <location filename="../../DlgAbout/DlgAbout.cpp" line="434"/>
         <source>Information</source>
         <translation>信息</translation>
     </message>